refactor: split two sizer about mode
