+++ package.json
@@ -12,7 +12,7 @@
     "tauri": "tauri",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
-    "test": "vitest"
+    "test": "vitest run && test-storybook"
   },
   "dependencies": {
     "@nextui-org/react": "^2.1.13",
@@ -34,6 +34,7 @@
     "@storybook/manager-api": "^7.5.2",
     "@storybook/react": "^7.5.2",
     "@storybook/react-vite": "^7.5.2",
+    "@storybook/test-runner": "^0.13.0",
     "@storybook/testing-library": "^0.2.2",
     "@tauri-apps/cli": "^1.5.0",
     "@types/react": "^18.2.15",
@@ -41,6 +42,7 @@
     "@types/uuid": "^9.0.6",
     "@vitejs/plugin-react": "^4.0.3",
     "autoprefixer": "^10.4.16",
+    "playwright": "^1.39.0",
     "postcss": "^8.4.31",
     "storybook": "^7.5.2",
     "tailwindcss": "^3.3.5",
