+++ src/component/sizer/index.ts
@@ -1,0 +1,2 @@
+export * from './simple';
+export * from './complicated';
