+++ src/component/sizer/simple.stories.tsx
@@ -1,32 +1,25 @@
 import { StoryObj, type Meta } from "@storybook/react";
-import { Sizer } from ".";
+import { SimpleSizer } from "./simple";
 import { useState } from '@storybook/preview-api';
 import { Value } from "./type";
 import { within, userEvent } from '@storybook/testing-library';
 import { expect } from '@storybook/jest'
 
 const meta = {
-  title: "内置组件/Sizer",
-  component: Sizer,
+  title: "内置组件/Simple-Sizer",
+  component: SimpleSizer,
   tags: ["autodocs"],
   parameters: {
     layout: "centered",
   },
-  argTypes: {
-    mode: {
-      control: 'select',
-      options: ['simple', 'complicated']
-    },
-  }
-} satisfies Meta<typeof Sizer>;
+} satisfies Meta<typeof SimpleSizer>;
 
 export default meta;
-type Story = StoryObj<Meta<typeof Sizer>>;
+type Story = StoryObj<Meta<typeof SimpleSizer>>;
 
 export const SimpleMode: Story = {
   args: {
     label: 'width',
-    mode: 'simple',
     value: 10
   },
   decorators: [
@@ -45,7 +38,7 @@ export const SimpleMode: Story = {
     const canvas = within(canvasElement);
     const unitSelector = canvas.getByRole('unit-selector');
 
-    await step('Simple ui', () => {
+    await step('Simple ui', async () => {
       const inputs = canvasElement.querySelectorAll('input');
       expect(inputs.length).toEqual(1);
       expect(inputs[0].value).toEqual('10');
@@ -59,7 +52,6 @@ export const SimpleMode: Story = {
       const inputs = canvasElement.querySelectorAll('input');
       expect(inputs[0].value).toEqual('');
       expect(inputs[0].disabled).toBeTruthy();
-
       await userEvent.selectOptions(unitSelector, 'px');
       expect(inputs[0].value).toEqual('10');
     })
@@ -69,7 +61,6 @@ export const SimpleMode: Story = {
 export const PercentUnit: Story = {
   args: {
     label: 'width',
-    mode: 'simple',
     value: '12%'
   },
   decorators: [
@@ -89,48 +80,7 @@ export const PercentUnit: Story = {
 export const DisabledSimple: Story = {
   args: {
     label: 'width',
-    mode: 'simple',
     value: 'auto'
-  },
-  decorators: [
-    (story, config) => {
-      const [value, setValue] = useState<Value>(config.args.value);
-      return story({
-        args: {
-          ...config.args,
-          value: value,
-          onChange: setValue,
-        },
-      });
-    },
-  ],
-}
-
-export const ComplicatedMode: Story = {
-  args: {
-    label: 'margin',
-    mode: 'complicated',
-    value: 10
-  },
-  decorators: [
-    (story, config) => {
-      const [value, setValue] = useState<Value>(config.args.value);
-      return story({
-        args: {
-          ...config.args,
-          value: value,
-          onChange: setValue,
-        },
-      });
-    },
-  ],
-}
-
-export const ComplicatedValue: Story = {
-  args: {
-    label: 'margin',
-    mode: 'complicated',
-    value: '10px 20px'
   },
   decorators: [
     (story, config) => {
